requirements.txt
flask==2.1.1
gym==0.21.0
pandas==1.3.0
waitress==2.1.1
jupyterlab>=3.4.0
tensorflow==1.13.1
torch==1.9.0
# optional: RLlib and WarpDrive for training
# For CPU
# ray[rllib]==1.0.0 
# For GPU
# rl-warp-drive==1.6.7
matplotlib==3.2.2
numpy==1.21.6
deepdiff==5.8.1
pyyaml==6.0